feat: support for basic auto assignment
When no users have been assigned and the game mode is classic, then you can auto assign players, this will be based on currently the highest mean skill set and lowest number of games played.

This is not defined to be final implementation but a basic simplistic first attempt. Maybe in the future this could be shifted to the server and more computed based assignment can be performed.